Add bool method
